requirements.txt
python-telegram-bot==21.6
gTTS==2.5.3
python-dotenv==1.0.1
